+++ src/infrastructure/prun-ui/tiles.ts
@@ -69,6 +69,7 @@ function activateFrame(frame: HTMLDivElement, anchor: HTMLDivElement) {
   const tileElement = frame.parentElement!;
   const container = tileElement.parentElement!;
   const docked = !container.classList.contains(C.Window.body);
+  const window = container.closest(`.${C.Window.window}`) as HTMLElement | null;
   const id = getPrunId(tileElement)!;
   const commandElement = _$(frame, C.TileFrame.cmd);
   const fullCommand = commandElement!.textContent!.trim();
@@ -78,6 +79,7 @@ function activateFrame(frame: HTMLDivElement, anchor: HTMLDivElement) {
     container,
     frame,
     anchor,
+    window,
     docked,
     fullCommand,
     command: (indexOfSpace > 0 ? fullCommand.slice(0, indexOfSpace) : fullCommand).toUpperCase(),