+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - `ADM`: The "Withdraw" vote button now has a gray color (`adm-neutral-withdraw-button`)
 - `CONT`: The preamble of user-made contracts is no longer minimized (`minimize-headers`)
 - `FXPO`: The Current Price field is now set to the order book price (`fxpo-better-current-price`)
+- `SFC`: The window now auto-closes on success (`sfc-auto-close`)
 - `XIT BURN`: Workforce consumables now include remaining fractional allocations
 - `XIT BURN`: The "Days" column now displays fractional days when less than 10 days remain
 - `XIT BURN`: The "Need" column is now based on fractional days instead of whole days